+++ pkg/wasm/cache_test.go
@@ -40,6 +40,7 @@ import (
 	"github.com/google/go-containerregistry/pkg/v1/types"
 
 	extensions "istio.io/api/extensions/v1alpha1"
+	"istio.io/istio/pkg/bootstrap/platform"
 	"istio.io/istio/pkg/util/sets"
 )
 
@@ -621,7 +622,7 @@ func TestWasmCache(t *testing.T) {
 			if c.wasmModuleExpiry != 0 {
 				options.ModuleExpiry = c.wasmModuleExpiry
 			}
-			cache := NewLocalFileCache(tmpDir, options)
+			cache := NewLocalFileCache(tmpDir, &platform.Unknown{}, options)
 			cache.httpFetcher.initialBackoff = time.Microsecond
 			defer close(cache.stopChan)
 
@@ -800,7 +801,7 @@ func setupOCIRegistry(t *testing.T, host string) (dockerImageDigest, invalidOCII
 
 func TestWasmCachePolicyChangesUsingHTTP(t *testing.T) {
 	tmpDir := t.TempDir()
-	cache := NewLocalFileCache(tmpDir, defaultOptions())
+	cache := NewLocalFileCache(tmpDir, &platform.Unknown{}, defaultOptions())
 	defer close(cache.stopChan)
 
 	gotNumRequest := 0
@@ -855,7 +856,7 @@ func TestAllInsecureServer(t *testing.T) {
 	tmpDir := t.TempDir()
 	options := defaultOptions()
 	options.InsecureRegistries = sets.New("*")
-	cache := NewLocalFileCache(tmpDir, options)
+	cache := NewLocalFileCache(tmpDir, &platform.Unknown{}, options)
 	defer close(cache.stopChan)
 
 	// Set up a fake registry for OCI images with TLS Server
