+++ pkg/wasm/keychain/bootstrap/bootstrap_test.go
@@ -1,0 +1,30 @@
+// Copyright Istio Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package bootstrap
+
+import (
+	"testing"
+
+	"istio.io/istio/pkg/wasm/keychain"
+)
+
+// Test the keychains are correctly loaded.
+func TestLoadingKeychain(t *testing.T) {
+	wantNumberOfKeychains := 4
+	gotNumberOfKeychains := keychain.GetRegisteredKeychainsCount()
+	if wantNumberOfKeychains != gotNumberOfKeychains {
+		t.Errorf("unexpected # of keychains (want: %d, got: %d)", wantNumberOfKeychains, gotNumberOfKeychains)
+	}
+}
